Assign releases to the right commit
Summary:
This was a tough one. It took about 20 builds and reading the github docs to
figure out wtf is going on.

1) Travis needs a tag to even start the deployment.
2) Adding a tag to the local Travis checkout works fine.
3) However, upon receipt of the Travis request to create a new release, Github,
not being aware of this tag (since Travis can't push it to github), defaults to
creating it on master. Wherever master is pointing right now.

The fix is to specify the `target_commitish` argument explicitly.
Separately, we create a better name for the tag - something like
`release-<shortrev>` rather than the mess travis generates.

Also, let's disable email notifications for successful builds while we're here.
Also-also, let's not start the build again once github pushes the tag, that's
silly.

Differential Revision:
D6456285
Ninja: oss build stuff

fbshipit-source-id: 3318d0c2a24ee21dc73d2ab0ce81b6b1896b5375