Prevent ANRs during Multi-Process tracing
Summary:
This diff prevents ANR situation when main process waits for a secondary process to finish the trace file.
The fix will improve stability of the app while we are tracing but it doesn't address the tracing logic behind the exceptional cases when ANR happens. This will be addressed separately.

I added checks for some exceptional situations in Multi-Process tracing:

1. Do not register trace in secondary process if `adoptContext` has failed.
2. Take into account `traceId` when stopping trace and do not wait for closure if match was not found.
3. Check if `traceId` was registered in the secondary process in `onTraceAbort`, `onTraceStop`

I was able to see all the 3 situations when tested.

Reviewed By: BurntBrunch

Differential Revision: D6775963

fbshipit-source-id: 00473f2587ebc671a98f5436590026c92f415419