Refactor all provider constants into a dynamic registry of names
Summary:
For the longest time, Loom has maintained a single list of providers. This has
become problematic in the face of open sourcing the library and the increasing
number of apps using it.

This diff creates a `ProvidersRegistry` which assigns a numeric identifier (a
bit in a bitset) for each call to `newProvider`.

The ability to create more than one bit for the same string is intentional -
it allows the user to split a single configuration string into multiple
distinct runtime values. In the future, we'll likely support mapping more
than one string to the same bit constant too, so we can effectively
split providers in a backward-compatible way.

In this world, config providers only ask the `ProvidersRegistry` for the
appropriate bits for their strings.

Reviewed By: aandreyeu

Differential Revision: D7008382

fbshipit-source-id: 5a0e1dc16654f08bbc7575a616c7275a8c060a12