Add template type T as friend class to allow exposure of newInstance without having to rewrite the functionality outside of the class
Summary: newInstance is a protected function of JavaClass. This means any class directly implementing JavaClass gets newInstance for free but because HybridClass has a JavaPart that implements JavaClass and is a friend of HybridClass it can be called from within HybridClass (i.e. newObjectCxxArgs) but anything that implements HybridClass cannot directly call this and if they want to pass arguments to a custom Java constructor for the object they must rewrite the functionality in newInstance in each class that wants to do so. This makes JavaPart a friend of the template type so that this function can be called directly.

Reviewed By: mhorowitz, ricardorey10

Differential Revision: D7279643

fbshipit-source-id: ecc43f1813371e37dd5edf38d08d767efde3ea8d