modify art headers for trampolines, logging.cc -> logging-inl.h
Summary:
Static globals need to be converted to methods so that our generated-code trampolines can look up and return the actual, real value.

The logging subsystem requires a particular class to be defined for sizing purposes, but this class lives in the logging.cc file - place it alone in a header and include that header in the generated-code trampoline translation unit so it's able to reason about the class size.

Reviewed By: andrurogerz

Differential Revision: D7156557

fbshipit-source-id: 182a855bef2466757f989c005f8816cee11c393a