Enable PLT hooks on Android N+
Summary:
Since Android N, Google changed the semantics of `dlopen()`, causing it to return a handle instead of a pointer to a `soinfo` struct. Also, using `dlopen()` on system libraries is not allowed, so we're in double trouble.

There's currently a solution for this in our linker library (parsing `/proc/self/maps`), but it's quite slow, so this diff introduces a new mechanism for installing hooks.

The new mechanism is based on `dl_iterate_phdr`. We basically call `dl_iterate_phdr` to iterate over every loaded shared object and, for each one of those shared objects, we stash:

- The relocation table for PLT entries (DT_JMPREL)
- The size of the relocation table
- The symbols table
- The strings table

With these, it is possible for us to find the address of the relocation of a symbol - in this case, of functions such as `open()`, `close()`, etc - and install our hooks.

Reviewed By: aandreyeu

Differential Revision: D6757893

fbshipit-source-id: 2e6cb644a37056a4d5d162944705dcf88c245f01