Issue 15: Account for missing VALUE traces
Summary: Traces in which a STRING_KEY does not have a corresponding STRING_VALUE fail to be parsed correctly. This scenario is somewhat uncommon but not impossible, as it can happen that a trace is stopped right in the window after the KEY is written but before the VALUE is written. As this is only an issue of a missing single point, it's better to silently skip it than to fail parsing the whole thing.

Reviewed By: aandreyeu

Differential Revision: D7455982

fbshipit-source-id: 451984f7f6167c6f4b8a2d98a3380d5ca9447bcf