add ability to no-op CHECK macros to prevent both sigabrt and deadlocks
Summary:
CHECK macros are not safe to no-op solely via `LogMessage.stream()` - some of them will loop indefinitely, and it's not always possible to avoid SIGABRT because <reasons??>

We also can't simply no-op redefine them, because the evaluation itself can have side effects (easily visible case: CHECK_PTHREAD_CALL), and also because the stream operators.

use macros and boolean logic instead to keep the test evaluation and macro body, but just force the actual test to false when we want.

Reviewed By: BurntBrunch

Differential Revision: D7410958

fbshipit-source-id: b54cf91ece703b6053ae1cceb46b83fa6d902387