Fix sizeof(art::Thread) mismatch on 7.0
Summary: We introduced some new members to the QuickEntrypoints list in order to get at various interpreter entrypoint functions, but that was unfortunately the wrong way to do it - the QuickEntrypoints list is an actual data structure that we don't get to dictate the layout of. So, revert the additions and allow access to the entrypoint functions a different way.

Reviewed By: aandreyeu

Differential Revision: D7596088

fbshipit-source-id: 934b87b9d9a0e3ae1ad193cd2def49d729c2841d