Handle case in device.py where files can move from one directory to another
Summary:
When we get a trace, it is created in the `cache/` directory, then moved to
`cache/upload`, where it stays until it is successfully uploaded (usually),
and finally it is taken back to 'cache/' after the upload is successful.

When we're looking for traces, we do a recursive "find" over the "default"
app's directory (i.e., the output of "adb shell run-as <package> pwd"), so it
can find things both in `cache/` and `cache/upload`. We store this as THE
file's path.

The problem with this is that, if the file moved from `cache/upload` from
`cache/`, and we cached that the file lived in `cache/upload`, then subsequent
commands will fail with "No such file or directory", leading to failure to
extract the trace.

The solution is to throw a particular exception if the file was moved, and to
retry for as long as necessary - while we're getting that exception -. The
possible scenarios are:

1) File is in `cache/` and it stays there the whole time. There's no problem here.

2) File is in `cache/upload` and it stays there the whole time. There's no problem here.

3) File was in `cache/` when we ran the script, but at some point after we
found it there it moved to `cache/upload`. This triggers a retry that will go
to /2/ or /5/.

4) File was in `cache/upload` when we ran the script, but at some point after
we found it there it was moved back to `cache/`. This triggers a retry that
will take us to /1/.

4) The file actually moved back to `cache/` before we finished.
This should trigger a retry that would take us to /1/.

Reviewed By: BurntBrunch, guohuixie

Differential Revision: D7854676

fbshipit-source-id: 430a07762d078e0ed27f8385c55ccf152f951e72