Build a custom trampoline for each hook
Summary:
Clang seems to implement PLT calls with nonreturning branches (`b` instead of `bl`, `jmp` instead of `call`), which breaks liblinker's `CALL_PREV` mechanism which relies on knowing exactly who the logical caller of the hook is.

So, install a custom-built trampoline as the hook, and have _it_ call the provided hook, setting up the stack and return values manually so we know for sure where we're coming from. Each installed hook will get its very own trampoline, so we can uniquely tell where the call came from.

This only fixes the problem for ARM; x86 and aarch64 will simply (gracefully) fail to install the trampoline.

Reviewed By: BurntBrunch

Differential Revision: D8125469

fbshipit-source-id: b1593898ec24d939c622d345e025a298f97127d8