External API improvement
Summary:
Due to linker changes in recent Android versions symbol lookup is restricted and current approach will fail for libraries outside ones which are merged together with the API.

The solution is to have a minimalistic shared library (external_api_glue) which can be dynamically linked from the customer libs. Profilo, when loaded initialized the API and set implementation function references making it active.

Reviewed By: BurntBrunch

Differential Revision: D8169289

fbshipit-source-id: 106b6a97cd937b228d7e7c1cef086753b81adeb0