GitHub issue 23
Summary:
`STRING_KEY` entries with a value `__name` should only have one child (namely, a STRING_VALUE entry) and there's an assert for that. This parent-child relationship is established by means of one of the `arg`s of the entry - the parent of a `StandardEntry` is its `arg2`, whereas the parent of a `BytesEntry` is its `arg1` -. The bug here was that `arg2` for `StandardEntry` entries of type `CPU_COUNTER`  is specialized to mean "core number" rather than "parent". In the log that was uploaded to github, multiple `CPU_COUNTER` entries had `3` as their `arg2`, which led the parser to incorrectly mark all of those entries as children of entry 3, which happened to be a `STRING_KEY` entry with value `__name`, and thus firing the assert.

The solution is simply to not add `CPU_COUNTER` entries to the children list, as the concept of parent-children is meaningless there.

Reviewed By: BurntBrunch

Differential Revision: D8271156

fbshipit-source-id: 123c674524b1edd86b269866ca43374b7e116cb3