Safer compat checks for museum tracer
Summary:
In extracting ArtCompatibilityRunner, I mistakenly replaced the `InitRuntime`
call with effectively `InstallRuntime`. The latter requires that we understand
the ABI but it runs outside of the fork jail in ArtCompatibilityRunner.

This diff introduces a new `prepare` method to handle this case. `prepare()` is
called outside of the jail, initializes statics and does symbol lookups.
`startTracing()` is now first called inside the jail to ensure it's safe.

Reviewed By: aandreyeu

Differential Revision: D8667444

fbshipit-source-id: dc4e97fe58fa41ffec2a91f9c43d16cbe118002c