Work around buggy art::FaultHandler
Summary:
`art::FaultManager` is faulty in the way described in the comment in
`sigmuxsetup.h`

We can work around this by ensuring that the sigmux handler is in front of
sigchain's. However, to make this stick, we also need to bump the target SDK to
something more realistic, as art will keep resetting itself in front if we
target something less than 21.

Reviewed By: aandreyeu

Differential Revision: D8527211

fbshipit-source-id: e2e67840a9bbdc5f8da7e3eb6607d71678c2d93c