update notebook package
Summary:
Jupyter < 5.4.1 is vulnerable to CVE-2018-8768:

```
In Jupyter Notebook before 5.4.1, a maliciously forged notebook file can bypass sanitization to execute JavaScript in the notebook context. Specifically, invalid HTML is 'fixed' by jQuery after sanitization, making it dangerous.
```

This is not really that big of a deal for us but there's no reason not to
update.

Reviewed By: aandreyeu

Differential Revision: D8830145

fbshipit-source-id: fe45ed341bc6a64f254016b450cc608e47660f09