Remove 5.1 museum unwinder
Summary:
In our haste to add 5.1 unwinding, we didn't check if the libc
pthread_getspecific implementation is async-signal-safe. Why wouldn't it be?

Because it's implemented as a shared table with exclusive locking :(

We need pthread_get/setspecific to be lock-free for double-entry prevention,
but also because that's how we find the `art::Thread` object.

Therefore, unwinding on 5.1 is not safe and should be taken out until we have a
better plan.

Reviewed By: aandreyeu

Differential Revision: D8866334

fbshipit-source-id: 6a37431262f8326575fe85f5b2855832ac688d07