Short-circuit isObjectRefType for null references
Summary: `isObjectRefType` is supposed to always return `true` when asked if `null` is a given reference type. In situations where the JNI environment is not yet attached, however (such as during global initialization), it will cause an assertion failure due to inability to check the JNI version. Since that's actually irrelevant for null references, perform the null check first and only check the version once we actually need to.

Reviewed By: ricardorey10

Differential Revision: D8882566

fbshipit-source-id: 1af1f463d87876af209370c6e11d5498c6925b6b