Log actually tracing providers
Summary:
The intent of this diff is to log the actual tracing state of providers in the trace. Currently every provider silently hides the tracing state and can be idle while there is no easy way to know it from the trace metadata.

I processed every provider class to add a capability to request if it's actually running. Infomation about actually tracing providers will be collected on trace end before disabling the providers.
Some of the providers doesn't require initialization (like `other`) in this case we suppose that such providers are tracing as long as they are enabled in the trace context.

Reviewed By: BurntBrunch

Differential Revision: D9376521

fbshipit-source-id: 0f077e738eddc68e16289e85eb54a7cd1edfa68d