Log triggers instead of providers during trace start
Summary:
During D9473303 I added code to annotate - in the QPL Metrics table - the name
of the trigger responsible of starting a trace. However, instead of logging
the trigger name, I was logging the provider's name, which is not the same.

This was due to me totally misunderstanding Aliaksei's comment about reusing
the name resolution in `GenericRegistry`, using the functionality he added
in `ProvidersRegistry` instead.

Reviewed By: BurntBrunch

Differential Revision: D9615661

fbshipit-source-id: 8c5f4799a049acbb3b7bfa2a76bda25d70654f35