Remove sample app code from aar
Summary:
I was a little lazy when I made the aars and just reused the sample app
machinery. This is problematic because sample apps have extra classes and a
bigger manifest.

This diff removes manifest entry for the sample app activity and strips out the
sample app code.

This fixes issue #32

Reviewed By: ricardorey10

Differential Revision: D9726548

fbshipit-source-id: d1c2278821d9aceaeab7e1b2400b854523d6bd89