Make counters log at a consistent timestamp
Summary:
We currently log most counter values at unique timestamps. This happens because
we don't capture a singular timestamp when reading the file but instead add it
when issuing the Logger call.

This diff fixes that by using the same timestamp for all counters that are part
of the same file.

Reviewed By: tmelz

Differential Revision: D9724125

fbshipit-source-id: e59f485ef8e988a769dcbb800b27faefa085404c