Avoid JNI roundtrips in TraceEvents
Summary:
There are many calls to `TraceEvents.isEnabled` in a well-instrumented app.
A lot of these calls happen regardless of whether we are currently tracing.

This diff ensures that we do not cross the JNI barrier for `isEnabled` calls.
This is achieved by having both `disable` and `enable` return the new mask.
That way we can cache it in Java and just check for bit presence in pure Java.

A requirement for correctness is that all changes happen through the Java
layer, which is currently the case.

Reviewed By: aandreyeu

Differential Revision: D9410574

fbshipit-source-id: 080a136df5d4b34df45830569d8967f5f58ef378