Fix js_tracer static linking
Summary:
`js_tracer` is mistakenly using static linking for `logger:logger`. This is a
problem as it messes up which copy `libprofilo_stacktrace` ends up using, which
in turn means that data ends up missing (since we now have two buffers but only
one of them is being read from).

Reviewed By: gavinweng

Differential Revision: D12829847

fbshipit-source-id: 3ef5daded6f2753894ac42c8250214d229869824