Implement PLT unhooking
Summary:
This diff introduces the opposite of PLT hooking - PLT unhooking.

In practice, this is effectively modifications of the run list for a given GOT
slot with the optional reporting of when only one item is left in the list
(the "trivial" case) and the undoing of the GOT slot modification itself.

Currently, trampolines are not deallocated as that's a much harder
concurrency problem.

Reviewed By: tophyr, ricardorey10

Differential Revision: D10023820

fbshipit-source-id: d615dbd22cdf0a695cde11f28e554472daff85a4