Change hooking methodology to use one trampoline per GOT slot
Summary:
The current implementation of GOT hooking does not track the hook list
explicitly. Instead, it stores the previous value of the GOT slot inside every
trampoline, effectively building a singly-linked list:

```
[ slot ] -> [trampoline 2] (.chained)-> [trampoline 1] (.chained)-> original
```

This diff rewrites the algorithm to instead store a "hook id" inside the
trampoline (as opposed to actual addresses of other hooks). This allows us to
maintain exactly one trampoline per GOT slot, with out-of-band hook lists
(called run lists in the code as they include the original).

```
[ slot ] -> [trampoline]
```

The contract between the trampoline and the rest of the library is therefore
changed as follows:

```
void* hook = push_hook_stack(hook_id, <return address>);
hook(<arguments>);
void* return_address = pop_hook_stack();
longjmp(return_address);
```

Further, the push/pop_hook_stack contract has been reimplemented to work a bit
differently. Instead of an explicit stack of hook entries, we now maintain a
singly-linked list of trampolines we have entered. Each entry for a trampoline
has a complete run list and an index into it. `get_chained_plt_method` also
advances (in practice, decrements) the index to point further toward the
original function.

Instead of this:

```
[ hook 1 for write ]
[ hook 2 for write ]
[ hook 3 for write ]
[ hook 1 for read ] (a read hook which calls write as well)
```

we now have this:

```
0x00: [ trampoline id 2 (write), run_list = (original, hook 1, hook 2, hook 3), run_idx = 0 , previous = 0xAA]
0xAA: [ trampoline id 1 (read), run_list = (original, hook 1), run_idx = 0, previous = 0x0]
```

In other words, `push` makes a new entry, copies the relevant run list, and puts the
previous TLS slot value in `.previous`. On the other hand, `pop` restores the value
of `.previous` into the TLS slot and releases the memory for the linked list item.

Reviewed By: tophyr

Differential Revision: D9998616

fbshipit-source-id: 4a7fca14c1fc3a36c0c50ecce531fea0186a8627