Correctly handle requests to unhook with unknown hooks
Summary:
There are two major cases where we need to indicate to the caller that the
unhook request wasn't actually wrong, just a no-op:

1) Slot is not hooked but matches the spec.
2) Slot is hooked and matches the spec but this hook function is not in the run list.

Previously, the code handled 1) correctly by checking `hooks::is_hooked` but
that was insufficient.

This diff changes the logic to handle both cases uniformly, via a new return
code. This also makes the locking a lot saner (the `is_hooked` call was
acquiring and releasing locks we would need shortly after).

Reviewed By: tophyr

Differential Revision: D12884294

fbshipit-source-id: 2419dda0157d2901d3a54cb68d664be59c715a81