Fix for CPU counter
Summary:
When `high_freq_main_thread_counters` provider is enabled `THREAD_CPU_TIME` is not logged.
The newly added `StatFileType:IO` wasn't a multiple of 2 and cleared the `StatFileType:SCHEDSTAT` from `availableStatFiles_` mask, in case where `/proc/../io` is not available.
The fix is to set proper constant for `IO`.

Reviewed By: BurntBrunch

Differential Revision: D12945280

fbshipit-source-id: 5ec9ffe2f54f7eda500e7110dd36afa953eb71ba