Switch 100% to /profilo subfolder in FileManager
Summary:
I noticed that my traces are created in `cache/` folder and sometimes both in `cache/` and `cache/profilo` folders.
I think we can safely remove migration portion now and switch 100% to `profilo` subfolder.
As having this duality seems to be error-prone. Every time a user clears the cache we end up in the migration behavior.

Reviewed By: BurntBrunch

Differential Revision: D13371068

fbshipit-source-id: cc45c01358f5b3eadb2164f986c119ee7d46e051