Use strong CAS when running final signal handler after exhausting registration list
Summary:
When sigmux receives an incoming signal, it runs through its `NORMAL` and then `LOW` priority registered handlers and if none of them indicated to continue execution (meaning that sigmux should continue searching for handlers) then it will, finally, run any "original" signal handler that was registered prior to sigmux's initialization.

If that original signal handler was marked as `SA_RESETHAND`, indicating that it should only ever be executed once, sigmux will atomically clear its storage of the original handler before executing it. It does this atomically to ensure that if two incoming signals are racing each other on separate threads, it still obeys the contract of "only executed once" and the loser of the CAS operation will not execute anything.

*However*, specifying a weak CAS operation means that there *doesn't have to be a winner* and that even after every thread contending for the update has completed its CAS, the value might never have been updated. Because we're using the CAS success to determine whether or not we should execute the original handler, this means that it's possible to not execute the handler at *all*. Sigmux needs to use a strong CAS to ensure that the handler will get executed exactly once.

Reviewed By: BurntBrunch

Differential Revision: D14188558

fbshipit-source-id: 8e17a4368e13b3d57113e672f55f5bbf527cff76