Quick providers
Summary: Adding a notion of "quick" provider which is initialized synchronously with a trace start. While current "normal" providers will init asynchronously on the worker thread as it's now. Quick providers list is essential for providers which are critical to init as soon as trace starts so we start capturing the data early.

Reviewed By: BurntBrunch

Differential Revision: D14733177

fbshipit-source-id: e91893ae9af5351b4308a877d26226b1855c4dcf