Fix broken build
Summary:
Ubuntu 17.10 is reaching End Of Life (EOL) according to https://wiki.ubuntu.com/Releases, and it looks like support is already dying, as the Travis build has been broken for a while.

This changes the Ubuntu version to 16.04, which will reach EOL in 2024, and makes the appropriate changes to make the `docker build` command work.

Reviewed By: tmelz, BurntBrunch, aandreyeu

Differential Revision: D15327841

fbshipit-source-id: 694481b967371b24c8193addeb0ca759c69148a7