Fail fast for disabled external tracer
Summary:
Whenever the SIGPROF handler is called, we get a slot (which involves some operations on atomics), iterate over all the tracers, and do a bunch of work before calling the `collectStack()` method for each one of them.

Sometimes, the `ExternalTracer` is disabled (i.e., the callback is null), so doing all of this is wasteful. This diff extends the `ExternalTracer` API to provide a way of knowing if it is disabled, so that we can fail faster.

My original idea was to completely remove disabled tracers from the `tracersMap`, but Delyan said it can happen that at some point in time a tracer is disabled but later it gets enabled, so we can't just remove it. Bummer :(

Reviewed By: BurntBrunch

Differential Revision: D15357498

fbshipit-source-id: f3897aa187fcf9a37690916294b1de6b32369b8c