Re-add 5.x tracers
Summary:
This diff re-adds the removed 5.1 tracer and adds a new 5.0 tracer.

5.1 support was originally based on museum but that's not why we removed it. We
removed it because `pthread_setspecific` is not lock-free before Android 6. We
needed it to be because we used TLS to avoid nested unwinds.

(Note that we still need `pthread_getspecific` to be lock-free and indeed it is).

Now that this is no longer a concern, we can reintroduce these tracers.

Reviewed By: ricardorey10

Differential Revision: D15505003

fbshipit-source-id: b585498fa7857f90abb2a9ef632c002aa98a109c