Introduce a new state, BUSY_WITH_METADATA, to allow reliable time access
Summary:
We are currently a bit unprincipled w.r.t when we take a timestamp in the
profiling handler.

This diff unifies the various places and introduces a consistent location - in
getSlotIndex. Further, we extend the slot state list with BUSY_WITH_METADATA,
which indicates that the timestamp (and potentially other metadata in the
future) has been populated.

The primary reason to do is to allow consistent timestamp access in the
fault handler.

Reviewed By: aandreyeu

Differential Revision: D15425252

fbshipit-source-id: 82bf80b93b7f00bcfb8c2d5db9a02a79fc85f8d4