Elminate temporary string allocations in fbsystrace provider
Summary:
Turns out there are a lot of temporary string allocations in the external api calls where the compiler automatically does new `std::string` allocations to convert C-strings to STL strings.

I'd like to avoid that unnecessary allocation and `std::string_view` would be an ideal solution, but we are not yet there with C++17. So I decided to go with integer constants exposed with static allocation on demand of their `std::string` equivalents.

Reviewed By: ricardorey10

Differential Revision: D15954152

fbshipit-source-id: 61c8b1f89ffe78a3fa68a20f7f0c987a0abd1382