Add a provider that dumps file-backed memory mappings
Summary:
This provider is necessary for both perfevents and, in the future, native
stacks to work properly.

We create a new encoding of MAPPING entry with a specific key-value pair
containing the mapping.

All numbers are in hex and the key tells you the format
("start:end:offset:file" aka "s:e:o:f").

Reviewed By: aandreyeu

Differential Revision: D15869865

fbshipit-source-id: 650bbf2e31d755849d125c2d7288e83cd87c209a