Fix thread ordering issues in perfevents
Summary:
This diff addresses 2 issues:

First, it was previously possible to start and stop a reader thread before it ever
began running. As a result clock measurement stalled forever on the `join()` of
the reader thread.

This diff fixes this issue by ensuring stop() waits for the reader to begin first.
It also renames things to be slightly more consistent between Reader and Session.

Second, this diff fixes the acess to reader_ in Session and ensures writes are
visible to the reader thread by using a lock.

Reviewed By: aandreyeu

Differential Revision: D15990350

fbshipit-source-id: 760939d3f5e6ed38c5d0aae242ed24df05784a4d