Filter out anonymous minor faults
Summary:
Around half of the data flowing through the `faults` provider is minor faults for anonymous memory ranges.
While interesting quantitatively, it's not super interesting that we touched a page of anonymous memory for the first time.

Therefore, to help with the event rate, we filter out all known anonymous memory ranges. We do this in two ways:

1. We use procmaps to snapshot the known mappings and only keep the non-anonymous mappings
2. We enable perfevents mmap records, which allow us to keep track of the mappings set in real time.

Together, this allows us to have an accurate view of the current file-backed mappings and thus we can ignore everything else.

Reviewed By: tmelz

Differential Revision: D16969452

fbshipit-source-id: a3fba8c754e23088fa4809e7c476da8dffa813fc