Remove support for android-arm
Summary:
android-arm is the old pre-armv7 ABI (armeabi), and it isn't supported by newer
NDKs. We currently use the r17 NDK, which doesn't even have any of the armeabi
libraries (only armeabi-v7a, corresponding to android-armv7), so it's not
buildable. Remove support for it from the build system, since all that support
is just dead code.

Reviewed By: asp2insp

Differential Revision: D17643117

fbshipit-source-id: fcde11a9469e019590c47fba9434e25838b6e2d1