Extract OrderedKeyedStatFile
Summary:
Simple refactor, fixing a shadowing bug for `read_` in the process.

We can no longer use references to fields in `stat_info_` because they are not initialized at the time of the reference construction. (Which is fine, it's just an address, but I don't want to suppress valuable warnings.) As a result we have to use pointer-to-member which ends up being like 2 armv7 instruction longer to read/write but whatever.

Reviewed By: ricardorey10, aandreyeu

Differential Revision: D18579817

fbshipit-source-id: 9eb059ff7e2bc9c099c786fc44d37f7ff44e6afe