Enable RTTI alongside exceptions and drop -fPIE
Summary:
In general, enabling exceptions without also enabling RTTI is a bad
idea. It causes the compiler to generate local copies of the typeinfos
for classes with key functions (whose typeinfo should in fact be
anchored inside the libraries defining the key function), because it
assumes that the libraries defining the key function must also have been
built without RTTI and it therefore can't rely on them to provide the
typeinfo object. These local typeinfo copies will cause failures for
strict (pointer-based) typeinfo comparisons, which could e.g. cause
exceptions to not be caught across libraries.

On Android, we use loose (string-based) typeinfo comparisons, but we
still run into problems with this. In order to avoid potential clashes
between the system libc++ on some phones and our custom libc++, we
perform linker wrapping for any libc++ symbols that don't have the
`__ndk1` inline namespace, which includes the typeinfo for various
classes (e.g. std::logic_error). If we enable exceptions without
enabling RTTI, the local typeinfo objects will be generated, and if
we're building with -fPIE (which as it turns out is the default for
64-bit Android targets), the compiler will generate relative relocations
against the local typeinfo copies (instead of GOT-relative relocations,
as it would with -fPIC). The linker wrapping will then replace
references to these typeinfos with references to the corresponding
wrapped symbols, which are external (they live inside libc++). Relative
relocations can't be used against these external symbols (they need
GOT-relative relocations), so we'll get a link error.

Fix this both by enabling RTTI, which prevents unnecessary local copies
of the typeinfo from being generated, and by removing any explicit
-fPIE flags, at which point Buck's default of -fPIC will kick in. -fPIC
is what we want for shared libraries anyway; -fPIE only works because we
also build with -Bsymbolic (but would e.g. break for references to local
data symbols if we were to build with -Bsymbolic-functions instead).

Unfortunately, it turns out enabling RTTI causes some problems of its
own (e.g. link failures because we expect to find external typeinfo that
isn't actually present because of the external library not building with
RTTI). We'll therefore globally enable -fPIC in a follow-up, although
it's still good to fix the RTTI/exceptions mismatch wherever we can.

Reviewed By: schwink

Differential Revision: D18774503

fbshipit-source-id: 77465bff4e59ffe2be9337fdc8ff3e9724bfad48