feat(website): add DocSearch as recommended by docusaurus. (#59)
Summary:
👋 team,

I am working on DocSearch. [We have an integration for Docusaurus](https://docusaurus.io/docs/en/search#enabling-the-search-bar). We thought it is a shame you can not also used this search that you can [find on reactjs for example](https://reactjs.org/).

This PR will add DocSearch to the documentation website. It will allow an user to have a learn-as-you-type experience by displaying results thanks to a dropdown in a live way.

Let me know if you need anything.
Pull Request resolved: https://github.com/facebookincubator/profilo/pull/59

Reviewed By: tmelz

Differential Revision: D20041561

Pulled By: yukonfb

fbshipit-source-id: 7cd295da97453849e22ea51178b23887bcba11b5