Make the fmt build agnostic of the library's version
Summary:
`deps/fmt/BUCK` hardcodes the version of the library (currently 5.2.0). This is bad, because whenever the library is updated in `third-party/fmt`, we'll copy a different version into our GitHub repo, causing the build to break. This is what's happening in D20198332.

This diff removes the hardcoded `5.2.0` from the BUCK file, so that this can keep building regardless of the library's version.

Reviewed By: BurntBrunch

Differential Revision: D20344375

fbshipit-source-id: 935b471f8f1febd794c554e7d916086f03143276