Tweak RESET_ORIG_SIGACTION_FLAG sigmux_reinit flag
Summary:
I consolidated the logic here to make a little bit more sense. Previously we weren't actually reallocating and then we are freeing so that isn't great.

I also made it so we redo the sigaction call for the default sigaction and capture the original sigaction like we do in sigmux_init

Reviewed By: stashuk

Differential Revision: D20478848

fbshipit-source-id: e54607c257ff33fd4dea65d71f41395b9a3f76bf