Fix longstanding round_up bug
Summary:
Looks like when a native method was the leaf we would get the frame_size incorrectly because our implementation of RoundUp would accidentally add 16 in the case that size is already a multiple of 16.

I noticed we ported the round_up function for support in other OS 7+ code, so I could simply re-use that function :).

Reviewed By: BurntBrunch

Differential Revision: D21263709

fbshipit-source-id: bb57d729d27d8eea5ee84c6fdf530eaedd34c046