Add support for EntryTypeEnum as its own type
Summary:
This diff introduces explicit support for the EntryType enum, as its own type
in the codegen. The actual codegen uses explicit casts to uint8_t internally,
so there's no change to how things are packed/unpacked.

There's also a check now that the number of entry types is < 255 (we're nowhere
close but still).

Reviewed By: aandreyeu

Differential Revision: D21330493

fbshipit-source-id: d238323f7918f93e2e9c46fa2cbe1548962f23c3