Less strict read count check
Summary:
The current check is too aggressive as it supposes we reach half of the buffer capacity when process dies.
However it may not be the case if it happens early enough. So I am loosing this check so we still collect the trace if it contains at least one readable entry.

Reviewed By: BurntBrunch

Differential Revision: D21385536

fbshipit-source-id: 1caf626e4ea8d11b24db9e314c20082e24234c87