Use -frtti to avoid typeinfo copies
Summary:
We're using `-fexceptions` universally but we're not using `-frtti`. This causes very weird bugs that libc++ exposes.

Namely, we end up with a copy of `typeinfo for std::system_error` within `libprofilo.so` and `libprofilo_util.so` and symbol resolution sometimes picks *that* up as opposed to the one from `libc++_shared.so`. Our libc++ does string comparisons for types, so this doesn't affect us but it does affect the libc++ bundled in the NDK.

Reviewed By: Fortisque

Differential Revision: D21416835

fbshipit-source-id: 08daffc4bbf14de5464dee75c54700d130cf250a